finalise content + move client to app dir
