+++ content/docsTableOfContents/YakShaver/toc.mdx
@@ -1,0 +1,53 @@
+---
+parentNavigationGroup:
+  - title: Introduction
+    items:
+      - title: Introduction to YakShaver
+        slug: content/docs/YakShaver/how-to-use-yakshaver.mdx
+        _template: item
+      - title: Tips to Creating the Perfect Work Item
+        slug: content/docs/YakShaver/yakshaver-tips.mdx
+        _template: item
+      - title: FAQ
+        slug: content/docs/YakShaver/frequently-asked-questions.mdx
+        _template: item
+  - title: Setup Guides
+    items:
+      - title: Third Party Recording Tools Setup
+        slug: content/docs/YakShaver/third-party-tools-setup.mdx
+        _template: item
+      - title: Setting up the Screenshot App (MacOS)
+        slug: content/docs/YakShaver/recording-work-item-screenshot-app.mdx
+        _template: item
+      - title: Setting up the Snipping tools (Windows)
+        slug: content/docs/YakShaver/recording-work-item-snipping-tool.mdx
+        _template: item
+      - title: Setting up Microsoft Teams
+        slug: content/docs/YakShaver/teams-setup.mdx
+        _template: item
+      - title: Setting up Other Recording Tools
+        slug: content/docs/YakShaver/setting-up-other-recording-tools.mdx
+        _template: item
+  - title: Usage Guides
+    items:
+      - title: Tips to Create the Perfect Work Item
+        slug: content/docs/YakShaver/yakshaver-tips.mdx
+        _template: item
+      - title: User Mention Guidelines in Item Descriptions
+        slug: >-
+          content/docs/YakShaver/user-mention-guidelines-in-work-item-descriptions.mdx
+        _template: item
+      - title: Connecting your GitHub Account
+        slug: content/docs/YakShaver/connect-your-github-account.mdx
+        _template: item
+      - title: Recording a Work Item with Browser Extension
+        slug: content/docs/YakShaver/recording-work-item-browserextension.mdx
+        _template: item
+      - title: Recording a Work Item with Snagit
+        slug: content/docs/YakShaver/recording-work-item-snagit.mdx
+        _template: item
+      - title: Recording a Work Item with Microsoft Teams
+        slug: content/docs/YakShaver/recording-work-item-teams.mdx
+        _template: item
+---
+
