--- content/docsTableOfContents/toc.mdx
@@ -1,5 +1,0 @@
----
-parentNavigationGroup:
-  - title: Introduction
----
-
